requirements.txt
Django==1.11.4
django-bootstrap3==9.1.0
django-datatables-view==1.14.0
django-filter==1.1.0
django-tables2==1.10.0
mysqlclient==1.3.10
pytz==2017.2
